fix(browser): writing properties
Upgrade browserify and use readable-stream 4.x.x
Workaround to https://github.com/whatwg/fetch/issues/1438

Fix #785
